Forum is nearly completed!

+++ html/forum.php
@@ -37,8 +37,12 @@ else {
 			echo "<br><br>";
 			echo "What's happening: " .$row['post_content'];
 			echo "<div class='text-right'>";
+			if ($_SESSION['loginEmail'] == $row['post_by']) {
 			echo "<a href='questionview.php?id=$id'>View</a> | <a href='reply.php?id=$id'>Reply</a> | <a href='delete.php?id=$id'>Delete</a>";
-			echo "</div>";
+			} else {
+			echo "<a href='questionview.php?id=$id'>View</a> | <a href='reply.php?id=$id'>Reply</a>";
+}
+echo "</div>";
 			echo "</div>";
 			echo "</div>";
 			echo "<br></a>";
